third_party/WebKit/LayoutTests/platform/win/external/wpt/user-timing/test_user_timing_mark_and_measure_exception_when_invoke_with_timing_attributes-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
FAIL performance.mark and performance.measure should throw if used with timing attribute values assert_throws: function "function () { window.performance.measure(timingAttributes[i]); }" did not throw
FAIL performance.mark and performance.measure should not throw if used with timing attribute values in workers Failed to execute 'mark' on 'WorkerPerformance': 'connectEnd' is part of the PerformanceTiming interface, and cannot be used as a mark name.
Harness: the test ran to completion.
